requirements.txt
numpy
pandas
torch
torchvision
torchaudio
absl-py
pytorch-lightning

